[AndroidTests] Cleanup.
