+++ MiraiNotes.Android.UiTests/Pages/BasePage.cs
@@ -1,20 +1,10 @@
 ﻿using NUnit.Framework;
 using System;
-using System.Drawing;
-using System.Linq;
-using Xamarin.UITest;
-using Xamarin.UITest.Queries;
 
 namespace MiraiNotes.Android.UiTests.Pages
 {
-    public abstract class BasePage<TPage>
+    public abstract class BasePage : BaseView
     {
-        protected IApp App => AppManager.App;
-
-        protected bool OnAndroid => AppManager.Platform == Platform.Android;
-
-        protected bool OniOS => AppManager.Platform == Platform.iOS;
-
         public abstract PlatformQuery Trait { get; }
 
         protected BasePage()
@@ -52,51 +42,5 @@ namespace MiraiNotes.Android.UiTests.Pages
 
             Assert.DoesNotThrow(() => App.WaitForNoElement(Trait.Current, timeout: timeout), message);
         }
-
-        public string GetHint(string fromClass = "TextInputLayout")
-        {
-            return App.Query(x => x.Class(fromClass).Invoke("getHint"))
-                .FirstOrDefault()
-                .ToString();
-        }
-
-        public bool IsDrawerOpen()
-        {
-            return App.Query(x => x.Class("NavigationView")).Any();
-        }
-
-        protected void OpenDrawer()
-        {
-            //App.SwipeLeftToRight(0.99, 1000, true);
-            App.Tap(x => x.Marked("Open"));
-        }
-
-        public void PressBackButton()
-        {
-            App.Back();
-        }
-
-        public Color GetColor(AppQuery baseQuery, int index = 0)
-        {
-            int colorValue = Convert.ToInt32(App.Query(x => baseQuery.Invoke("getCurrentTextColor"))[index]);
-
-            var color = Color.FromArgb(colorValue);
-            return color;
-        }
-
-        public bool ColorsAreClose(Color a, Color b, int threshold = 500)
-        {
-
-            var rDist = Math.Abs(a.R - b.R);
-            var gDist = Math.Abs(a.G - b.G);
-            var bDist = Math.Abs(a.B - b.B);
-
-            if (rDist + gDist + bDist > threshold)
-                return false;
-
-            return true;
-        }
-
-        public abstract TPage OpenDrawer(bool open = true);
     }
 }
