+++ MiraiNotes.Android.UiTests/Pages/BaseView.cs
@@ -1,0 +1,57 @@
+﻿using System;
+using System.Drawing;
+using System.Linq;
+using Xamarin.UITest;
+using Xamarin.UITest.Queries;
+
+namespace MiraiNotes.Android.UiTests.Pages
+{
+    public class BaseView
+    {
+        public const string TaskRecyclerViewId = "TaskRecyclerView";
+        public const string MvxListViewClass = "MvxListView";
+        public const string MaterialButtonClass = "MaterialButton";
+
+        protected IApp App => AppManager.App;
+        protected bool OnAndroid => AppManager.Platform == Platform.Android;
+        protected bool OniOS => AppManager.Platform == Platform.iOS;
+
+        public bool IsDrawerOpen()
+        {
+            return App.Query(x => x.Class("NavigationView")).Any();
+        }
+
+        public void PressBackButton()
+        {
+            App.Back();
+        }
+
+        public string GetHint(string fromClass = "TextInputLayout")
+        {
+            return App.Query(x => x.Class(fromClass).Invoke("getHint"))
+                .FirstOrDefault()
+                .ToString();
+        }
+
+        public Color GetColor(AppQuery baseQuery, int index = 0)
+        {
+            int colorValue = Convert.ToInt32(App.Query(x => baseQuery.Invoke("getCurrentTextColor"))[index]);
+
+            var color = Color.FromArgb(colorValue);
+            return color;
+        }
+
+        public bool ColorsAreClose(Color a, Color b, int threshold = 500)
+        {
+
+            var rDist = Math.Abs(a.R - b.R);
+            var gDist = Math.Abs(a.G - b.G);
+            var bDist = Math.Abs(a.B - b.B);
+
+            if (rDist + gDist + bDist > threshold)
+                return false;
+
+            return true;
+        }
+    }
+}
